+++ Project_SWENG/Assets/Scripts/Level_Tile/Dice/Dice.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c85af37348be1ae4f980a2d809fc3139
+guid: 87bf4fac6647a7e4982ba05ed663af9a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
